Move VariableList to a separate file
